Use a modal progress dialog when reloading data.
This effectively keeps the same "synchronous loading" property of doing the load in the UI thread, but gives better feedback to the user. True asynchronous loading would require a pretty big rethink of the data model.

This exposes a pre-existing issue where data is loaded twice when changing mod or Factorio directories.