Fix uncaught exception if a graph has no beacons in use.
The "Show Graph Summary" feature would fail entirely if your graph does
not use any beacons. We now implement an operator bool on the
BeaconQualityPair struct and check it directly since the type should be
responsible for knowing whether it's in a valid state or not.

All callsites that previously initialized the struct with null
parameters simply call a separate constructor with a documentory
comment.