Avoid data exporter crash with research-progress products
Factorio 2.0 added the "research-progress" product type for recipes,
which is missing many of the fields which "item" and "fluid" has. It
might be possible/useful to treat these as pseudo-items to allow
building graphs to create them, but for now we should just ignore them
to avoid the data exporter crashing.

Fixes #130